Create a vector of bounding boxes

Bounding boxes denote the exten of geometries. It follows that bounding boxes are only defined for geometries that are located in space and have finite extent. Thus, vectors, lines, directions, rays, etc. does not have bounding boxes, while e.g. spheres, circles, segments, etc has. Since the extent of geometries cannot always be given exact (e.g. for circles), bounding boxes are defined in regular floating point precision. Bounding boxes can be compared for equality and be tested for whether they overlap. Adding bounding boxes together will give the bounding box containing both.

bbox(...)

is_bbox(x)

as_bbox(x)

is_overlapping(x, y)

x %overlaps% y

Arguments

...
Either a vector of geometries or a range of numeric vectors (4 for 2D bounding boxes and 6 for 3D) interpreted in the order xmin, ymin, zmin, xmax, ymax, zmax.

x, y
vectors of bounding boxes or geometries

Value

An euclid_bbox vector

Examples

# Construction
bbox(10, -2, 15, 0)

#> <Vector of bounding boxes in 2 dimensions>
#> [1] <<10, -2>, <15, 0>>

seg <- segment(point(sample(10, 4), sample(10, 4)),
               point(sample(10, 4), sample(10, 4)))

boxes <- bbox(seg)
boxes

#> <Vector of bounding boxes in 2 dimensions>
#> [1] <<5, 2>, <7, 4>>   <<5, 6>, <8, 7>>   <<6, 1>, <7, 3>>   <<10, 5>, <10, 6>>

# Comparison
boxes[1] == boxes

#> [1]  TRUE FALSE FALSE FALSE

boxes[1:2] %overlaps% boxes[3:4]

#> [1]  TRUE FALSE

# Addition
boxes[1] + boxes[2]

#> <Vector of bounding boxes in 2 dimensions>
#> [1] <<5, 2>, <8, 7>>

cumsum(boxes)

#> <Vector of bounding boxes in 2 dimensions>
#> [1] <<5, 2>, <7, 4>>  <<5, 2>, <8, 7>>  <<5, 1>, <8, 7>>  <<5, 1>, <10, 7>>

# Conversion
as.matrix(boxes)

#>      [,1] [,2] [,3] [,4]
#> [1,]    5    2    7    4
#> [2,]    5    6    8    7
#> [3,]    6    1    7    3
#> [4,]   10    5   10    6