Vector of circles

This function create circles from various combinations of input.

circle(...)

is_circle(x)

as_circle(x)

Arguments

...
Various input. See the Constructor section.

x
A circle vector or an object to convert to it

Value

An euclid_circle vector

Constructors

2 dimensional circles

Providing one point and one numeric vector will construct circles centered at the point with the radius given by the numeric.

Providing two point vectors will construct circles centered between the two points with a radius of half the distance between the two points.

Providing three point vectors will construct the unique circle that pass through the three points.

Examples

## 2 Dimensions

point1 <- point(runif(5), runif(5))
point2 <- point(runif(5), runif(5))
point3 <- point(runif(5), runif(5))
number <- exact_numeric(1:5)

# Construction with center and radius
circle(point1, number)

#> <Vector of circles in 2 dimensions>
#> [1] <x0:0.0808, y0:0.466, r:1> <x0:0.834, y0:0.498, r:2> 
#> [3] <x0:0.601, y0:0.29, r:3>   <x0:0.157, y0:0.733, r:4> 
#> [5] <x0:0.0074, y0:0.773, r:5>

# integers and numerics are converted automatically
circle(point1, 1:5)

#> <Vector of circles in 2 dimensions>
#> [1] <x0:0.0808, y0:0.466, r:1> <x0:0.834, y0:0.498, r:2> 
#> [3] <x0:0.601, y0:0.29, r:3>   <x0:0.157, y0:0.733, r:4> 
#> [5] <x0:0.0074, y0:0.773, r:5>

# You are free to name the input for readability
circle(center = point1, radius = number)

#> <Vector of circles in 2 dimensions>
#> [1] <x0:0.0808, y0:0.466, r:1> <x0:0.834, y0:0.498, r:2> 
#> [3] <x0:0.601, y0:0.29, r:3>   <x0:0.157, y0:0.733, r:4> 
#> [5] <x0:0.0074, y0:0.773, r:5>

# Construction with 2 points
circle(point1, point2)

#> <Vector of circles in 2 dimensions>
#> [1] <x0:0.478, y0:0.331, r:0.419> <x0:0.505, y0:0.451, r:0.333>
#> [3] <x0:0.318, y0:0.177, r:0.305> <x0:0.239, y0:0.561, r:0.19> 
#> [5] <x0:0.205, y0:0.874, r:0.222>

# Construction with 3 points
circle(point1, point2, point3)

#> <Vector of circles in 2 dimensions>
#> [1] <x0:0.491, y0:0.371, r:0.422>  <x0:0.518, y0:0.358, r:0.346> 
#> [3] <x0:0.0968, y0:0.73, r:0.669>  <x0:-0.144, y0:0.379, r:0.464>
#> [5] <x0:0.463, y0:0.372, r:0.607>