Access the exact values that defines the geometries

This function gives access to the underlying values defining the geometries in a vector. As such they return the same information as calling as.matrix() on a geometry vector except the return value is kept as an exact numeric and that you can extract from single elements if the cardinality of the geometry exceeds 1.

parameter(x, ...)

# S3 method for euclid_geometry
parameter(x, which, element = NA, ...)

# S3 method for euclid_affine_transformation
parameter(x, i, j, ...)

Arguments

x
A geometry vector

...
parameters to pass on

which
Either a name or the index of the definition to extract, as matched to definition_names(x)

element
For geometries with a cardinality above 1, which element of the geometry should the definition be extracted for. If NA the definition for all elements will be returned and the length of the returned vector will be sum(cardinality(x)) (matching the return of as.matrix(x))

i, j
The row and column of the cell in the transformation to fetch.

Value

An exact_numeric vector