Vector of points

Points are the fundamental unit in geometry, from which other geometric primitives can be constructed. Points support less than and greater than operations which are handled lexicographically. This means that point vectors can also be sorted and ranked. Subtracking a point from a point gives a vector. Adding and subtracting a vector to a point translates the point by the vector.

point(..., default_dim = 2)

is_point(x)

as_point(x)

Arguments

...
Various input. See the Constructor section.

default_dim
The dimensionality when constructing an empty vector

x
A point vector or an object to convert to it

Value

a euclid_point vector

Constructors

2 dimensional points

Providing two numeric vector will construct points with those x and y coordinate.

Providing a vector will construct a point at the location the vector points to.

Providing a weighted point vector will create a points at the same locations without weights.

3 dimensional points

Providing three numeric vector will construct points with those x, y, and z coordinate.

Providing a vector will construct a point at the location the vector points to.

Providing a weighted point vector will create a points at the same locations without weights.

Examples

num1 <- exact_numeric(runif(5))
num2 <- exact_numeric(runif(5))
num3 <- exact_numeric(runif(5))

# 2 dimensions
p <- point(num1, num2)
p

#> <2D points [5]>
#> [1] <x:0.167, y:0.379>  <x:0.0287, y:0.174> <x:0.786, y:0.604> 
#> [4] <x:0.825, y:0.805>  <x:0.965, y:0.0367>

# 3 dimensions
point(num1, num2, num3)

#> <3D points [5]>
#> [1] <x:0.167, y:0.379, z:0.733>  <x:0.0287, y:0.174, z:0.215>
#> [3] <x:0.786, y:0.604, z:0.016>  <x:0.825, y:0.805, z:0.129> 
#> [5] <x:0.965, y:0.0367, z:0.686>

# Standard R vectors are automatically converted to exact_numeric vectors
point(runif(5), runif(5))

#> <2D points [5]>
#> [1] <x:0.642, y:0.926> <x:0.327, y:0.459> <x:0.387, y:0.596> <x:0.708, y:0.165>
#> [5] <x:0.698, y:0.567>

# Convert points to vectors
as_vec(p)

#> <2D vectors [5]>
#> [1] <x:0.167, y:0.379>  <x:0.0287, y:0.174> <x:0.786, y:0.604> 
#> [4] <x:0.825, y:0.805>  <x:0.965, y:0.0367>

# Arithmetic
# Translate by adding vector
p + vec(3, 7)

#> <2D points [5]>
#> [1] <x:3.17, y:7.38> <x:3.03, y:7.17> <x:3.79, y:7.6>  <x:3.83, y:7.81>
#> [5] <x:3.97, y:7.04>

# Create vector by subtracting points
p[1:2] - p[3:4]

#> <2D vectors [2]>
#> [1] <x:-0.619, y:-0.225> <x:-0.797, y:-0.631>

# Sorting etc.
sort(p)

#> <2D points [5]>
#> [1] <x:0.0287, y:0.174> <x:0.167, y:0.379>  <x:0.786, y:0.604> 
#> [4] <x:0.825, y:0.805>  <x:0.965, y:0.0367>

min(p)

#> <2D points [1]>
#> [1] <x:0.0287, y:0.174>

cummax(p)

#> <2D points [5]>
#> [1] <x:0.167, y:0.379>  <x:0.0287, y:0.174> <x:0.0287, y:0.174>
#> [4] <x:0.0287, y:0.174> <x:0.0287, y:0.174>

rank(p)

#> [1] 2 1 3 4 5

p[1:2] < p[3:4]

#> [1] TRUE TRUE