Project geometries to lines and planes

Projecting geometries to lines and planes transforms the geometries to lie on the given line or plane by movement orthogonal to the line or plane. The geometry of the result of the projection is the same as the input. Not all geometries can be projected to lines and planes. Specifically, iso rectangles and iso cubes cannot do either and only 3 dimensional geometries can be projected to a plane. Circles and spheres are special in the sense that no ellipse and ellipsoid primitives exist. Because of this, projections will retain the radius and circularity of the primitives despite this not being geometrically sound. Projecting a 3D circle to a line will keep the orientation of the circle, while projecting it to a plane will adopt the orientation of the plane.

project(x, target)

Arguments

x
A vector of geometries to project or map

target
A vector of lines or planes

Value

A vector of geometries of the same class as x

Examples

# Project a collection of segments to a line
p <- point(sample(10), sample(10))
s <- segment(p[1:5], p[6:10])

project(s, line(3, 1, -6))

#> <Vector of segments in 2 dimensions>
#> [1] [<x:0.6, y:4.2>, <x:1, y:3>]      [<x:1.1, y:2.7>, <x:1.6, y:1.2>] 
#> [3] [<x:-0.7, y:8.1>, <x:1.6, y:1.2>] [<x:0.2, y:5.4>, <x:-0.2, y:6.6>]
#> [5] [<x:1.2, y:2.4>, <x:0.6, y:4.2>] 

# Project a line to the x plane
project(line(point(3, 7, -3), point(-2, 6, 0)),
        plane(point(0, 0, 0), vec(1, 0, 0)))

#> <Vector of lines in 3 dimensions>
#> [1] <x:0, y:7, z:-3, dx:0, dy:-1, dz:3>