Vector of vectors

A geometrical vector is somewhat different from the concept of a vector in programming hence the slightly confusing terminology. In geometry a vector is a direction and a magnitude most often defined by a point in space where the direction is defined as the direction from the origin to the point and the magnitude is defined as the distance from the origin to the point.

vec(...)

is_vec(x)

as_vec(x)

Arguments

...
Various input. See the Constructor section.

x
A vector of vectors or an object to convert to it

Value

An euclid_vector vector

Constructors

2 dimensional vectors

Providing a point will construct vectors pointing to the points from the origin centered.

Providing two exact numeric vectors will construct vectors pointing to the point defined by the coordinates given.

3 dimensional vectors

Providing a point will construct vectors pointing to the points from the origin centered.

Providing three exact numeric vectors will construct vectors pointing to the point defined by the coordinates given.