Vector of circles

A circle is defined by a center and a radius (given as the squared radius due to the inexactness of square roots). In 3 dimensions a circle is a disc and thus have an orientation given by an orthogonal direction. If the radius is 0 the circle is considered degenerate.

circle(..., default_dim = 2)

is_circle(x)

as_circle(x)

Arguments

...
Various input. See the Constructor section.

default_dim
The dimensionality when constructing an empty vector

x
A circle vector or an object to convert to it

Value

An euclid_circle vector

Constructors

2 dimensional circles

Providing one point and one numeric vector will construct circles centered at the point with the squared radius given by the numeric.

Providing two points will construct circles centered between the two points with a radius of half the distance between the two points.

Providing three point vectors will construct the unique circle that pass through the three points.

3 dimensional circles

Providing three point vectors will construct the unique circle that pass through the three points.

Providing a point, a numeric, and a plane will construct a circle centered on the point with the squared radius given by the numeric and the orientation given by the plane. The point must lie on the plane

Providing a point, a numeric, and a vector will construct a circle centered on the point with the squared radius given by the numeric and the orientation orthogonal to the vector

Providing two spheres will construct a circle given by the intersection of the two spheres. The spheres must intersect

Providing a sphere and a plane will construct a circle given by the intersection of the sphere and the plane. The sphere and plane must intersect.

Examples

## 2 Dimensions

point1 <- point(runif(5), runif(5))
point2 <- point(runif(5), runif(5))
point3 <- point(runif(5), runif(5))
number <- exact_numeric(1:5)

# Construction with center and radius
circle(point1, number)

#> <Vector of circles in 2 dimensions>
#> [1] <x:0.196, y:0.696, r2:1>  <x:0.981, y:0.689, r2:2> 
#> [3] <x:0.742, y:0.0312, r2:3> <x:0.0514, y:0.226, r2:4>
#> [5] <x:0.53, y:0.301, r2:5>  

# integers and numerics are converted automatically
circle(point1, 1:5)

#> <Vector of circles in 2 dimensions>
#> [1] <x:0.196, y:0.696, r2:1>  <x:0.981, y:0.689, r2:2> 
#> [3] <x:0.742, y:0.0312, r2:3> <x:0.0514, y:0.226, r2:4>
#> [5] <x:0.53, y:0.301, r2:5>  

# You are free to name the input for readability
circle(center = point1, radius = number)

#> <Vector of circles in 2 dimensions>
#> [1] <x:0.196, y:0.696, r2:1>  <x:0.981, y:0.689, r2:2> 
#> [3] <x:0.742, y:0.0312, r2:3> <x:0.0514, y:0.226, r2:4>
#> [5] <x:0.53, y:0.301, r2:5>  

# Construction with 2 points
circle(point1, point2)

#> <Vector of circles in 2 dimensions>
#> [1] <x:0.416, y:0.438, r2:0.115>  <x:0.73, y:0.453, r2:0.118>  
#> [3] <x:0.587, y:0.356, r2:0.129>  <x:0.379, y:0.362, r2:0.126> 
#> [5] <x:0.739, y:0.471, r2:0.0728>

# Construction with 3 points
circle(point1, point2, point3)

#> <Vector of circles in 2 dimensions>
#> [1] <x:0.579, y:0.578, r2:0.161> <x:0.476, y:0.722, r2:0.256>
#> [3] <x:0.399, y:0.266, r2:0.173> <x:0.408, y:0.293, r2:0.132>
#> [5] <x:0.803, y:0.393, r2:0.083>

## 3 Dimensions

point1 <- point(runif(5), runif(5), runif(5))
point2 <- point(runif(5), runif(5), runif(5))
point3 <- point(runif(5), runif(5), runif(5))

circ <- circle(point1, point2, point3)

circle(point1, number, as_vec(point2))

#> <Vector of circles in 3 dimensions>
#> [1] <x:0.532, y:0.0652, z:0.336, r2:1, dx:0.545, dy:0.466, dz:0.857>
#> [2] <x:0.494, y:0.354, z:0.596, r2:2, dx:0.279, dy:0.39, dz:0.385>  
#> [3] <x:0.779, y:0.825, z:0.192, r2:3, dx:0.447, dy:0.0201, dz:0.528>
#> [4] <x:0.204, y:0.274, z:0.948, r2:4, dx:0.372, dy:0.377, dz:0.601> 
#> [5] <x:0.713, y:0.57, z:0.542, r2:5, dx:0.0281, dy:0.56, dz:0.261>  

# Conversion
as_plane(circ)

#> <Vector of euclid_planes in 3 dimensions>
#> [1] <a:-0.112, b:-0.133, c:0.105, d:0.0329>     
#> [2] <a:-0.0581, b:0.0291, c:0.0641, d:-0.0198>  
#> [3] <a:-0.0979, b:0.0639, c:0.0561, d:0.0128>   
#> [4] <a:-0.0847, b:-0.00117, c:-0.0412, d:0.0566>
#> [5] <a:-0.0174, b:0.341, c:0.0302, d:-0.198>    

as_sphere(circ)

#> <Vector of euclid_spheres in 3 dimensions>
#> [1] <x:0.293, y:0.397, z:0.502, r2:0.194> 
#> [2] <x:0.317, y:0.159, z:0.525, r2:0.0742>
#> [3] <x:1.05, y:0.516, z:1.02, r2:0.847>   
#> [4] <x:0.302, y:0.211, z:0.747, r2:0.0537>
#> [5] <x:0.339, y:0.557, z:0.479, r2:0.144>