Vector of points

Points are the fundamental unit in geometry, from which other geometric primitives can be constructed. Points support less than and greater than operations which are handled lexicographically. This means that point vectors can also be sorted and ranked. Subtracking a point from a point gives a vector. Adding and subtracting a vector to a point translates the point by the vector.

point(..., default_dim = 2)

is_point(x)

as_point(x)

Arguments

...
Various input. See the Constructor section.

default_dim
The dimensionality when constructing an empty vector

x
A point vector or an object to convert to it

Value

a euclid_point vector

Constructors

2 dimensional points

Providing two numeric vector will construct points with those x and y coordinate.

Providing a vector will construct a point at the location the vector points to.

Providing a weighted point vector will create a points at the same locations without weights.

3 dimensional points

Providing three numeric vector will construct points with those x, y, and z coordinate.

Providing a vector will construct a point at the location the vector points to.

Providing a weighted point vector will create a points at the same locations without weights.

Examples

num1 <- exact_numeric(runif(5))
num2 <- exact_numeric(runif(5))
num3 <- exact_numeric(runif(5))

# 2 dimensions
p <- point(num1, num2)
p

#> <Vector of points in 2 dimensions>
#> [1] <x:0.296, y:0.763>  <x:0.984, y:0.417>  <x:0.59, y:0.138>  
#> [4] <x:0.759, y:0.0808> <x:0.836, y:0.656> 

# 3 dimensions
point(num1, num2, num3)

#> <Vector of points in 3 dimensions>
#> [1] <x:0.296, y:0.763, z:0.602>  <x:0.984, y:0.417, z:0.657> 
#> [3] <x:0.59, y:0.138, z:0.329>   <x:0.759, y:0.0808, z:0.979>
#> [5] <x:0.836, y:0.656, z:0.715> 

# Standard R vectors are automatically converted to exact_numeric vectors
point(runif(5), runif(5))

#> <Vector of points in 2 dimensions>
#> [1] <x:0.873, y:0.0461> <x:0.983, y:0.617>  <x:0.219, y:0.598> 
#> [4] <x:0.665, y:0.407>  <x:0.39, y:0.858>  

# Convert points to vectors
as_vec(p)

#> <Vector of vectors in 2 dimensions>
#> [1] <x:0.296, y:0.763>  <x:0.984, y:0.417>  <x:0.59, y:0.138>  
#> [4] <x:0.759, y:0.0808> <x:0.836, y:0.656> 

# Arithmetic
# Translate by adding vector
p + vec(3, 7)

#> <Vector of points in 2 dimensions>
#> [1] <x:3.3, y:7.76>  <x:3.98, y:7.42> <x:3.59, y:7.14> <x:3.76, y:7.08>
#> [5] <x:3.84, y:7.66>

# Create vector by subtracting points
p[1:2] - p[3:4]

#> <Vector of vectors in 2 dimensions>
#> [1] <x:-0.294, y:0.625> <x:0.224, y:0.336> 

# Sorting etc.
sort(p)

#> <Vector of points in 2 dimensions>
#> [1] <x:0.296, y:0.763>  <x:0.59, y:0.138>   <x:0.759, y:0.0808>
#> [4] <x:0.836, y:0.656>  <x:0.984, y:0.417> 

min(p)

#> <Vector of points in 2 dimensions>
#> [1] <x:0.296, y:0.763>

cummax(p)

#> <Vector of points in 2 dimensions>
#> [1] <x:0.296, y:0.763> <x:0.296, y:0.763> <x:0.296, y:0.763> <x:0.296, y:0.763>
#> [5] <x:0.296, y:0.763>

rank(p)

#> [1] 1 5 2 3 4

p[1:2] < p[3:4]

#> [1]  TRUE FALSE