Vector of points

Points are the fundamental unit in geometry, from which other geometric primitives can be constructed. Points support less than and greater than operations which are handled lexicographically. This means that point vectors can also be sorted and ranked. Subtracking a point from a point gives a vector. Adding and subtracting a vector to a point translates the point by the vector.

point(...)

is_point(x)

as_point(x)

Arguments

...
Various input. See the Constructor section.

x
A point vector or an object to convert to it

Value

a euclid_point vector

Examples

num1 <- exact_numeric(runif(5))
num2 <- exact_numeric(runif(5))
num3 <- exact_numeric(runif(5))

# 2 dimensions
point(num1, num2)

#> <Vector of points in 2 dimensions>
#> [1] <x:0.68, y:0.766>  <x:0.499, y:0.77>  <x:0.642, y:0.991> <x:0.66, y:0.971> 
#> [5] <x:0.096, y:0.389>

# 3 dimensions
point(num1, num2, num3)

#> <Vector of points in 3 dimensions>
#> [1] <x:0.68, y:0.766, z:0.461>  <x:0.499, y:0.77, z:0.315> 
#> [3] <x:0.642, y:0.991, z:0.175> <x:0.66, y:0.971, z:0.532> 
#> [5] <x:0.096, y:0.389, z:0.494>

# Standard R vectors are automatically converted to exact_numeric vectors
point(runif(5), runif(5))

#> <Vector of points in 2 dimensions>
#> [1] <x:0.779, y:0.825>  <x:0.204, y:0.274>  <x:0.713, y:0.57>  
#> [4] <x:0.0652, y:0.336> <x:0.354, y:0.596>